Lots of changes: 1. Function forwarder is now only used to figure out what to do with graphics objects that have no parent module. All other functionality is explicitly defined in the each module 2. Input text boxes now have an explicitly associated input toggle button 3. Input toggle buttons still have an explicitly associated input text box, however, the input text boxes with which they are associated are now explicitly defined on initialization, rather than determined via a poorly written function 4. All modules avoid accessing the audio in other modules directly, and instead they simply call a function that stores all necessary values locally, and then use those values. This is likely terribly inefficient, and I will address this issue later when implementing my K rate callback function for parameter updates 5. Many small bug fixes in the process functions of the modules Overall, this commit should increase stability and result in fewer confusing crashes.

+++ Modules/Sah.hpp
@@ -47,7 +47,7 @@ enum Sah_Dependencies
 
 class Sah: public Module
 {
-    float sample;
+    double sample;
     double time_to_next_sample;
 
     public:
